[LF-56]fix: 타이머 컴포넌트 분리로 불필요한 렌더링 최적화
